[DateTimePicker][Android] Fix double press bug after dismissing DateTimePicker dialog
On Android, dismissing the DateTimePicker dialog by clicking Cancel or by clicking on the background overlay will cause a bug whereby you have to press the DateTimePicker input twice into to launch the DateTimePicker dialog again. This is because the expandable overlay is still in the expanded state after dismissing the dialog. This commit fixes this bug by toggling the expandable overlay when the DateTimePicker dialog is dismissed.